second

+++ app/Http/Controllers/OrganisateurController.php
@@ -14,11 +14,14 @@ class OrganisateurController extends Controller
      */
     public function index()
     {
-            $organisateurs = users::join('organisateurs', 'organisateurs.users_id', '=', 'userss.id')
-            ->get();
-        return view('admin.page.organisateur.index', compact('organisateurs'));
-       
+        if (organisateurPermission() == false || auth()->user()->usertype == 1) {
 
+            $organisateurs = users::join('organisateurs', 'organisateurs.user_id', '=', 'users.id')
+                ->get();
+            return view('admin.page.organisateur.index', compact('organisateurs'));
+        } else {
+            return abort(401);
+        }
     }
 
 
@@ -27,8 +30,12 @@ class OrganisateurController extends Controller
      */
     public function create()
     {
-        //
-        return view('admin.page.organisateur.create');
+        if (organisateurPermission() == false || auth()->user()->usertype == 1) {
+
+            return view('admin.page.organisateur.create');
+        } else {
+            return abort(401);
+        }
     }
 
     /**
@@ -37,29 +44,32 @@ class OrganisateurController extends Controller
     public function store(Request $request)
     {
 
+        if (organisateurPermission() == true || auth()->user()->usertype == 1) {
+            $users = users::create([
+                'last_name' => $request->last_name,
+                'first_name' => $request->first_name,
+                'adresse' => $request->adresse,
+                'email' => $request->email,
+                'password' => $request->password,
+                'disponible' => $request->disponible,
+                'compagnie' => $request->compagnie,
+                'adresse_compagnie' => $request->adresse_compagnie,
+                'type_evenement_organiser' => $request->type_evenement_organiser,
+                'experience' => $request->experience,
+            ]);
 
-        $users = users::create([
-            'last_name' => $request->last_name,
-            'first_name' => $request->first_name,
-            'adresse' => $request->adresse,
-            'email' => $request->email,
-            'password' => $request->password,
-            'disponible' => $request->disponible,
-            'compagnie' => $request->compagnie,
-            'adresse_compagnie' => $request->adresse_compagnie,
-            'type_evenement_organiser' => $request->type_evenement_organiser,
-            'experience' => $request->experience,
-        ]);
-
-        $users = Organisateur::create([
-            'users_id' => $users->id,
-            'disponible' => $request->disponible,
-            'compagnie' => $request->compagnie,
-            'adresse_compagnie' => $request->adresse_compagnie,
-            'type_evenement_organiser' => $request->type_evenement_organiser,
-            'experience' => $request->experience,
-        ]);
-        return redirect()->route('organisateur.index')->with('success', 'Organisateur créé avec succès');
+            $users = Organisateur::create([
+                'users_id' => $users->id,
+                'disponible' => $request->disponible,
+                'compagnie' => $request->compagnie,
+                'adresse_compagnie' => $request->adresse_compagnie,
+                'type_evenement_organiser' => $request->type_evenement_organiser,
+                'experience' => $request->experience,
+            ]);
+            return redirect()->route('organisateur.index')->with('success', 'Organisateur créé avec succès');
+        } else {
+            return view('frontend.page.index');
+        }
     }
 
 
@@ -102,7 +112,11 @@ class OrganisateurController extends Controller
      */
     public function show(Organisateur $organisateur)
     {
-        return view('organisateur.show', compact('organisateur'));
+        if (organisateurPermission() == true || auth()->user()->usertype == 1) {
+            return view('organisateur.show', compact('organisateur'));
+        } else {
+            return view('frontend.page.index');
+        }
     }
 
     /**
@@ -111,7 +125,11 @@ class OrganisateurController extends Controller
     public function edit(Organisateur $organisateur)
     {
         //
-        return view('admin.page.organisateur.edit', compact('organisateur'));
+        if (organisateurPermission() == true || auth()->user()->usertype == 1) {
+            return view('admin.page.organisateur.edit', compact('organisateur'));
+        } else {
+            return view('frontend.page.index');
+        }
     }
 
     /**
@@ -119,36 +137,39 @@ class OrganisateurController extends Controller
      */
     public function update(Request $request, Organisateur $organisateur)
     {
+        if (organisateurPermission() == true || auth()->user()->usertype == 1) {
+            $request->validate([
+                'last_name' => 'required|string',
+                'first_name' => 'required|string',
+                'email' => 'required|email|unique:organisateurs,user_email',
+                'adresse' => 'nullable|string',
+                'disponible' => 'nullable|string',
+                'password' => 'required|string',
+                'compagnie' => 'required|string',
+                'adresse_compagnie' => 'required|string',
+                'type_evenement_organiser' => 'required|string',
+                'experience' => 'required|integer',
 
-        $request->validate([
-            'last_name' => 'required|string',
-            'first_name' => 'required|string',
-            'email' => 'required|email|unique:organisateurs,email,' . $organisateur->id,
-            'adresse' => 'nullable|string',
-            'disponible' => 'nullable|string',
-            'password' => 'required|string',
-            'compagnie' => 'required|string',
-            'adresse_compagnie' => 'required|string',
-            'type_evenement_organiser' => 'required|string',
-            'experience' => 'required|integer',
+            ]);
 
-        ]);
+            $organisateur->users()->update([
+                'last_name' => $request->last_name,
+                'first_name' => $request->first_name,
+                'adresse' => $request->adresse,
+                'email' => $request->email,
+                'password' => $request->password,
+                'disponible' => $request->disponible,
+                'compagnie' => $request->compagnie,
+                'adresse_compagnie' => $request->adresse_compagnie,
+                'type_evenement_organiser' => $request->type_evenement_organiser,
+                'experience' => $request->experience,
 
-        $organisateur->users()->update([
-            'last_name' => $request->last_name,
-            'first_name' => $request->first_name,
-            'adresse' => $request->adresse,
-            'email' => $request->email,
-            'password' => $request->password,
-            'disponible' => $request->disponible,
-            'compagnie' => $request->compagnie,
-            'adresse_compagnie' => $request->adresse_compagnie,
-            'type_evenement_organiser' => $request->type_evenement_organiser,
-            'experience' => $request->experience,
+            ]);
 
-        ]);
-
-        return redirect()->route('organisateur.index')->with('success', 'Organisateur mis à jour avec succès');
+            return redirect()->route('organisateur.index')->with('success', 'Organisateur mis à jour avec succès');
+        } else {
+            return view('frontend.page.index');
+        }
     }
 
     /**
@@ -156,8 +177,11 @@ class OrganisateurController extends Controller
      */
     public function destroy(Organisateur $organisateur)
     {
-        //
-        $organisateur->delete();
-        return redirect()->back()->with('success', '  a été supprimé avec succès.');
+        if (organisateurPermission() == true || auth()->user()->usertype == 1) {
+            $organisateur->delete();
+            return redirect()->back()->with('success', '  a été supprimé avec succès.');
+        } else {
+            return view('frontend.page.index');
+        }
     }
 }
